refactor(logistics): move SLA and time-window domain modules (M02b)
Move time-window into the Logistics domain and split SLA breach into a pure domain core plus a transitional DB adapter. Align the public barrel, anchored tests and context documentation without changing routes, queries, schema or public contracts.
